Add support for asyncpg
