+++ pymeos/pymeos/main/ttext.py
@@ -230,7 +230,7 @@ class TText(Temporal[str, 'TText', 'TTextInst', 'TTextSeq', 'TTextSeqSet'], ABC)
         return ttext_out(self._inner)
 
     @staticmethod
-    def read_from_cursor(value, _):
+    def read_from_cursor(value, _=None):
         if not value:
             return None
         if value[0] != '{' and value[0] != '[' and value[0] != '(':
