+++ pymeos/pymeos/time/timestampset.py
@@ -397,7 +397,7 @@ class TimestampSet:
         raise TypeError(f'Operation not supported with type {other.__class__}')
 
     @staticmethod
-    def read_from_cursor(value, _):
+    def read_from_cursor(value, _=None):
         if not value:
             return None
         return TimestampSet(string=value)
